requirements.txt
asgiref==3.2.3
Babel==2.7.0
Django==3.0
django-phonenumber-field==4.0.0
phonenumbers==8.11.0
pytz==2019.3
sqlparse==0.3.0
